requirements.txt
numpy>=1.24
pandas>=2.0
PyYAML>=6.0
pytest>=7.0
hypothesis>=6.0
